Move function to different files
