+++ rush00/ft_putchar.c
@@ -1,0 +1,6 @@
+#include <unistd.h>
+
+void	ft_putchar(char c)
+{
+	write(1, &c, 1);
+}
